integration: improve xdefi mocks
